--- src/contexts/UserContext.tsx
@@ -1,58 +1,0 @@
-import { createContext, ReactNode, useContext, useState } from "react";
-import { useNotificationContext } from "./NotificationContext";
-import { useCallback, useMemo } from "../@lib";
-
-interface User {
-  id: number;
-  name: string;
-  email: string;
-}
-
-interface userContextType {
-  user: User | null;
-  login: (email: string, password: string) => void;
-  logout: () => void;
-}
-
-const UserContext = createContext<userContextType | undefined>(undefined);
-
-export const UserProvider = ({ children }: { children: ReactNode }) => {
-  const [user, setUser] = useState<User | null>(null);
-  const { addNotification } = useNotificationContext();
-
-  const login = useCallback(
-    (email: string) => {
-      setUser({ id: 1, name: "홍길동", email });
-      addNotification("성공적으로 로그인되었습니다", "success");
-    },
-    [addNotification],
-  );
-
-  const logout = useCallback(() => {
-    setUser(null);
-    addNotification("로그아웃되었습니다", "info");
-  }, [addNotification]);
-
-  const userContextValue: userContextType = useMemo(() => {
-    return {
-      user,
-      login,
-      logout,
-    };
-  }, [user, login, logout]);
-
-  return (
-    <UserContext.Provider value={userContextValue}>
-      {children}
-    </UserContext.Provider>
-  );
-};
-
-// eslint-disable-next-line react-refresh/only-export-components
-export const useUserContext = () => {
-  const context = useContext(UserContext);
-  if (context === undefined) {
-    throw new Error("useUserContext must be used within an UserProvider");
-  }
-  return context;
-};
