+++ src/contexts/theme/ThemeContext.tsx
@@ -1,26 +1,15 @@
-import { createContext, ReactNode, useContext, useState } from "react";
-import { useCallback, useMemo } from "../@lib";
-
-interface themeContextType {
-  theme: string;
-  toggleTheme: () => void;
-}
+import { createContext, ReactNode, useContext } from "react";
+import { themeContextType } from "./ThemeTypes";
+import { useThemeStore } from "./ThemeStore";
+import { useMemo } from "../../@lib";
 
 const ThemeContext = createContext<themeContextType | undefined>(undefined);
 
 export const ThemeProvider = ({ children }: { children: ReactNode }) => {
-  const [theme, setTheme] = useState("light");
+  console.log("ThemeProvider");
+  const store = useThemeStore();
 
-  const toggleTheme = useCallback(() => {
-    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
-  }, []);
-
-  const themeContextValue: themeContextType = useMemo(() => {
-    return {
-      theme,
-      toggleTheme,
-    };
-  }, [theme, toggleTheme]);
+  const themeContextValue = useMemo(() => store, [store]);
 
   return (
     <ThemeContext.Provider value={themeContextValue}>
@@ -31,6 +20,7 @@ export const ThemeProvider = ({ children }: { children: ReactNode }) => {
 
 // eslint-disable-next-line react-refresh/only-export-components
 export const useThemeContext = () => {
+  console.log("useThemeContext");
   const context = useContext(ThemeContext);
   if (context === undefined) {
     throw new Error("useThemeContext must be used within an ThemeProvider");
