+++ src/contexts/theme/ThemeTypes.ts
@@ -1,0 +1,4 @@
+export interface themeContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
