+++ src/contexts/user/UserContext.tsx
@@ -1,0 +1,24 @@
+import { createContext, ReactNode, useContext } from "react";
+import { userContextType } from "./UserTypes";
+import { useUserStore } from "./UserStore";
+
+const UserContext = createContext<userContextType | undefined>(undefined);
+
+export const UserProvider = ({ children }: { children: ReactNode }) => {
+  const userContextValue = useUserStore();
+
+  return (
+    <UserContext.Provider value={userContextValue}>
+      {children}
+    </UserContext.Provider>
+  );
+};
+
+// eslint-disable-next-line react-refresh/only-export-components
+export const useUserContext = () => {
+  const context = useContext(UserContext);
+  if (context === undefined) {
+    throw new Error("useUserContext must be used within an UserProvider");
+  }
+  return context;
+};
