+++ src/contexts/user/UserTypes.ts
@@ -1,0 +1,11 @@
+export interface User {
+  id: number;
+  name: string;
+  email: string;
+}
+
+export interface userContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
